Added CreateConroller and EditController, from which all create and edit controller extend, except the tour controllers, because the tour differs too much from the other edit and create controllers
