+++ src/Presentation/EventSchedule/CreateHofkapelle.fxml
@@ -74,7 +74,7 @@
                 <TableColumn fx:id="selectedMusicalWorks" prefWidth="249.0" text="Title" />
               </columns>
             </TableView>
-            <Button fx:id="btnEditDetails" mnemonicParsing="false" onAction="#editHofkapelleInstrumentation" onMouseClicked="#editHofkapelleInstrumentation" text="Edit Details ..." GridPane.columnIndex="1" GridPane.rowIndex="7" />
+            <Button fx:id="btnEditDetails" mnemonicParsing="false" onAction="#editInstrumentation" text="Edit Details ..." GridPane.columnIndex="1" GridPane.rowIndex="7" />
          </children>
       </GridPane>
       <GridPane prefHeight="73.0" prefWidth="390.0" GridPane.rowIndex="3">
@@ -87,7 +87,7 @@
         </rowConstraints>
          <children>
             <Button id="discard" fx:id="btnCancel" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#cancel" prefHeight="25.0" prefWidth="80.0" text="Cancel" />
-            <Button fx:id="btnSaveNewHofkapelle" mnemonicParsing="false" onAction="#insertNewHofkapellePerformance" prefHeight="25.0" prefWidth="80.0" text="Save" GridPane.columnIndex="1" />
+            <Button fx:id="btnSaveNewHofkapelle" mnemonicParsing="false" onAction="#insertEventDuty" prefHeight="25.0" prefWidth="80.0" text="Save" GridPane.columnIndex="1" />
          </children>
       </GridPane>
       <GridPane GridPane.rowIndex="2">
