+++ src/Presentation/EventSchedule/CreateOperaController.java
@@ -1,288 +1,26 @@
 package Presentation.EventSchedule;
 
-import Application.DTO.EventDutyDTO;
-import Application.DTO.InstrumentationDTO;
-import Application.DTO.MusicalWorkDTO;
-import Application.EventScheduleManager;
-import Utils.Enum.EventStatus;
 import Utils.Enum.EventType;
-import Utils.DateHelper;
-import Utils.MessageHelper;
-import Utils.PlanchesterConstants;
-import Utils.PlanchesterMessages;
-import com.jfoenix.controls.JFXDatePicker;
-import com.jfoenix.controls.JFXTimePicker;
-import javafx.beans.property.ReadOnlyStringWrapper;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.control.*;
-import javafx.stage.Stage;
-import javafx.stage.WindowEvent;
+
 import javax.xml.bind.ValidationException;
-import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.LinkedList;
-import java.util.List;
-
-
 
 /**
  * Created by Ina on 08.04.2017.
  */
-public class CreateOperaController {
-    @FXML private TextField name;
-    @FXML private TextArea description;
-    @FXML private JFXTimePicker startTime;
-    @FXML private JFXTimePicker endTime;
-    @FXML private JFXDatePicker date;
-    @FXML private TextField eventLocation;
-    @FXML private TextField conductor;
-    @FXML private TextField points;
-    @FXML private Button editDetails;
+public class CreateOperaController extends CreateController{
 
-    @FXML private TableView<String> musicalWorkTable;
-    @FXML private TableColumn<String, String> selectedMusicalWorks;
-
-    private MusicalWorkDTO musicalWork;
-    private InstrumentationDTO instrumentation; // TODO timebox2
-
-    public static List<EventDutyDTO> rehearsalList;
-    @FXML private TableView<String> rehearsalTableView;
-    @FXML private TableColumn<String, String> rehearsalTableColumn;
-
-
-
+    @Override
     @FXML
-    public void initialize() {
-        initializeMandatoryFields();
-        rehearsalList = new LinkedList<>();
-        selectedMusicalWorks.setCellValueFactory(param -> new ReadOnlyStringWrapper(param.getValue()));
-        points.textProperty().addListener((observable, oldValue, newValue) -> {
-            //^\d*\.\d{2}$
-            //"^\\d*[\\.,]?\\d{1,2}?$"
-            if (!newValue.matches("\\d*[\\,.]?\\d*?")) {
-                points.setText(newValue.replaceAll("[^\\d*[\\,.]?\\d*?]", ""));
-            }
-        });
+    public void editInstrumentation() {
+        super.setMusicalWorkMulitpleSelection(false);
+        super.editInstrumentation();
     }
 
-
+    @Override
     @FXML
-    private void insertNewOperaPerformance() throws ValidationException {
-        if(validate()) {
-            EventDutyDTO eventDutyDTO = new EventDutyDTO();
-            eventDutyDTO.setEventDutyID(null);
-            eventDutyDTO.setName(name.getText());
-            eventDutyDTO.setDescription(description.getText());
-            eventDutyDTO.setStartTime(DateHelper.mergeDateAndTime(date.getValue(), startTime.getValue()));
-            eventDutyDTO.setEndTime(endTime.getValue() == null ? DateHelper.mergeDateAndTime(date.getValue(), startTime.getValue().plusHours(2)) : DateHelper.mergeDateAndTime(date.getValue(), endTime.getValue()));
-            eventDutyDTO.setEventType(EventType.Opera);
-            eventDutyDTO.setEventStatus(EventStatus.Unpublished);
-            eventDutyDTO.setConductor(conductor.getText());
-            eventDutyDTO.setEventLocation(eventLocation.getText());
-
-            if(musicalWork != null) {
-                List<MusicalWorkDTO> selectedMusicalWorks = new ArrayList<MusicalWorkDTO>();
-                selectedMusicalWorks.add(musicalWork);
-                eventDutyDTO.setMusicalWorks(selectedMusicalWorks);
-            } else {
-                eventDutyDTO.setMusicalWorks(null);
-            }
-
-            eventDutyDTO.setPoints(((points.getText() == null || points.getText().isEmpty()) ? null : Double.valueOf(points.getText())));
-            eventDutyDTO.setInstrumentation(null); //alternative instrumentation is not in the first timebox
-            eventDutyDTO.setRehearsalFor(null);
-
-            eventDutyDTO = EventScheduleManager.createEventDuty(eventDutyDTO);
-            EventScheduleController.addEventDutyToGUI(eventDutyDTO); // add event to agenda
-            EventDutyDTO eventDutyDTO1 = EventScheduleManager.getEventDutyByDetails(eventDutyDTO);
-
-            for(EventDutyDTO eventD : rehearsalList){
-                eventD.setRehearsalFor(eventDutyDTO1.getEventDutyID());
-                eventD = EventScheduleManager.createEventDuty(eventD);
-                EventScheduleController.addEventDutyToGUI(eventD);
-            }
-
-            rehearsalList.clear();
-
-            EventScheduleController.setDisplayedLocalDateTime(eventDutyDTO.getStartTime().toLocalDateTime()); // set agenda view to week of created event
-            EventScheduleController.resetSideContent(); // remove content of sidebar
-            EventScheduleController.setSelectedAppointment(eventDutyDTO); // select created appointment
-        }
-    }
-
-    
-    @FXML
-    public boolean cancel() {
-        if(!name.getText().isEmpty() || !description.getText().isEmpty() || date.getValue() != null
-                || !eventLocation.getText().isEmpty() || !conductor.getText().isEmpty() || !points.getText().isEmpty() || musicalWork != null) {
-            ButtonType answer = MessageHelper.showConfirmationMessage(PlanchesterMessages.DISCARD_CHANGES);
-            if(ButtonType.NO.equals(answer)) {
-                return false;
-            }
-        }
-        // remove content of sidebar
-        EventScheduleController.resetSideContent();
-        return true;
-    }
-
-    @FXML
-    public void editOperaInstrumentation() {
-        InstrumentationController.selectMultipleMusicalWorks = false;
-        if(date.getValue() != null) {
-            InstrumentationController.newHeading = name.getText() + " | " + date.getValue().toString();
-        } else {
-            InstrumentationController.newHeading = name.getText();
-        }
-
-        InstrumentationController.selectedMusicalWorks = new ArrayList<>();
-        if(musicalWork != null) {
-            InstrumentationController.selectedMusicalWorks = new ArrayList<>();
-            InstrumentationController.selectedMusicalWorks.add(musicalWork);
-        }
-
-        FXMLLoader fxmlLoader = new FXMLLoader();
-        fxmlLoader.setLocation(getClass().getResource("Instrumentation.fxml"));
-        Scene scene = null;
-        try {
-            scene = new Scene(fxmlLoader.load());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        Stage stage = new Stage();
-        stage.setTitle("Musical Work & Instrumentation");
-        stage.setScene(scene);
-        stage.setOnCloseRequest(we -> {
-            if(InstrumentationController.apply) {
-                if(!InstrumentationController.selectedMusicalWorks.isEmpty()) {
-                    musicalWork = InstrumentationController.selectedMusicalWorks.get(0);
-                    musicalWorkTable.getItems().clear();
-                    musicalWorkTable.getItems().add(musicalWork.getName());
-                }  else {
-                    musicalWorkTable.getItems().clear();
-                    musicalWork = null;
-                }
-                // TODO: timbox 2 save instrumentation
-            }
-
-        });
-        InstrumentationController.stage = stage;
-
-        stage.showAndWait();
-    }
-  
-    @FXML
-    public void addNewRehearsal() {
-        FXMLLoader fxmlLoader = new FXMLLoader();
-        fxmlLoader.setLocation(getClass().getResource("CreateRehearsal.fxml"));
-        Scene scene = null;
-        try {
-            scene = new Scene(fxmlLoader.load());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        Stage stage = new Stage();
-        stage.setTitle("Add new Rehearsal");
-        stage.setScene(scene);
-        stage.show();
-
-        stage.setOnCloseRequest(we -> {
-            if(CreateRehearsalController.apply) {
-                rehearsalList.add(CreateRehearsalController.eventDutyDTO);
-                rehearsalTableView.getItems().clear();
-                rehearsalTableColumn.setCellValueFactory(param -> new ReadOnlyStringWrapper(param.getValue()));
-                for(EventDutyDTO e : rehearsalList) {
-                    String rehearsalToAdd = e.getName();
-                    rehearsalTableView.getItems().add(rehearsalToAdd);
-                }
-            }
-        });
-
-        CreateRehearsalController.stage = stage;
-
-    }
-
-    @FXML
-    public void removeRehearsal() {
-        String rehearsalToRemove = rehearsalTableView.getSelectionModel().getSelectedItem();
-        for(EventDutyDTO eventDutyDTO : rehearsalList) {
-            if(eventDutyDTO.getName().equals(rehearsalToRemove)) {
-                rehearsalList.remove(eventDutyDTO);
-                break;
-            }
-        }
-        rehearsalTableView.getItems().remove(rehearsalTableView.getSelectionModel().getFocusedIndex());
-        rehearsalTableColumn.setCellValueFactory(param -> new ReadOnlyStringWrapper(param.getValue()));
-        for(EventDutyDTO e : rehearsalList) {
-            String rehearsalToAdd = e.getName();
-            rehearsalTableView.getItems().add(rehearsalToAdd);
-        }
-    }
-
-    private void initializeMandatoryFields() {
-        name.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-        date.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-        startTime.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-
-        name.textProperty().addListener((observable, oldValue, newValue) -> {
-            if(name.getText() == null || name.getText().isEmpty()) {
-                name.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                name.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-        date.valueProperty().addListener((observable, oldValue, newValue) -> {
-            if(date.getValue() == null) {
-                date.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                date.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-
-        startTime.valueProperty().addListener((observable, oldValue, newValue) -> {
-            if(startTime.getValue() == null) {
-                startTime.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                startTime.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-    }
-
-    private boolean validate() {
-        LocalDate today = LocalDate.now();
-        LocalTime start = startTime.getValue();
-        LocalTime end = endTime.getValue();
-
-        if(name.getText().isEmpty()){
-            MessageHelper.showErrorAlertMessage("The Name is missing.");
-            name.requestFocus();
-            return false;
-        } else if(date.getValue() == null || date.getValue().isBefore(today) ){
-            MessageHelper.showErrorAlertMessage("The date is not valid.");
-            date.requestFocus();
-            return false;
-        } else if(start == null) {
-            MessageHelper.showErrorAlertMessage("The starttime is missing.");
-            return false;
-        } else if(end != null && (start.isAfter(end) || start.equals(end))) {
-            MessageHelper.showErrorAlertMessage("The endtime is not after the starttime. ");
-            return false;
-        } else if(date.getValue().equals(today) && start.isBefore(LocalTime.now())){
-            MessageHelper.showErrorAlertMessage("The starttime must be in future.");
-            date.requestFocus();
-            return false;
-        } else if(musicalWork == null){
-            MessageHelper.showErrorAlertMessage("A musical work has to be selected.");
-            editDetails.requestFocus();
-            return false;
-        }
-        return true;
+    protected void insertEventDuty() throws ValidationException {
+        super.setEventType(EventType.Opera);
+        super.insertEventDuty();
     }
 }
