+++ src/Presentation/EventSchedule/CreateRehearsalController.java
@@ -17,7 +17,7 @@ import java.time.LocalTime;
 /**
  * Created by Christina on 28.04.2017.
  */
-public class CreateRehearsalController {
+public class CreateRehearsalController extends CreateController {
 
     @FXML private TextField name;
     @FXML private TextArea description;
@@ -32,21 +32,22 @@ public class CreateRehearsalController {
     public static Stage stage;
     public static EventDutyDTO eventDutyDTO;
 
-
+    @Override
     @FXML public void initialize() {
         stage = new Stage();
-        initializeMandatoryFields();
+        super.initializeMandatoryFields();
     }
 
+    @Override
     @FXML
-    void cancel() {
+    public boolean cancel() {
         stage.fireEvent( new WindowEvent(stage, WindowEvent.WINDOW_CLOSE_REQUEST));
+        return true;
     }
 
     @FXML
     void applyNewRehearsal() {
         if(validate()) {
-
             eventDutyDTO = new EventDutyDTO();
             eventDutyDTO.setName(name.getText());
             eventDutyDTO.setDescription(description.getText());
@@ -64,36 +65,8 @@ public class CreateRehearsalController {
         }
     }
 
-    private void initializeMandatoryFields() {
-        name.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-        date.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-        startTime.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-
-        name.textProperty().addListener((observable, oldValue, newValue) -> {
-            if(name.getText() == null || name.getText().isEmpty()) {
-                name.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                name.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-        date.valueProperty().addListener((observable, oldValue, newValue) -> {
-            if(date.getValue() == null) {
-                date.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                date.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-
-        startTime.valueProperty().addListener((observable, oldValue, newValue) -> {
-            if(startTime.getValue() == null) {
-                startTime.setStyle(PlanchesterConstants.INPUTFIELD_MANDATORY);
-            } else {
-                startTime.setStyle(PlanchesterConstants.INPUTFIELD_VALID);
-            }
-        });
-    }
-
-    private boolean validate() {
+    @Override
+    protected boolean validate() {
         LocalDate today = LocalDate.now();
         LocalTime start = startTime.getValue();
         LocalTime end = endTime.getValue();
