+++ src/Presentation/EventSchedule/EditHofkapelle.fxml
@@ -66,7 +66,7 @@
                 <TableColumn fx:id="selectedMusicalWorks" prefWidth="239.0" text="Title" />
               </columns>
             </TableView>
-            <Button fx:id="btnEditDetails" mnemonicParsing="false" onAction="#editHofkapelleInstrumentation" text="Edit Details ..." GridPane.columnIndex="1" GridPane.rowIndex="7" />
+            <Button fx:id="btnEditDetails" mnemonicParsing="false" onAction="#editInstrumentation" text="Edit Details ..." GridPane.columnIndex="1" GridPane.rowIndex="7" />
             </children>
             <padding>
                 <Insets left="10.0" right="10.0" />
